First crack at renumberable opcodes for `serialize_and_run_program`.
