+++ src/py/f_table.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use crate::allocator::Allocator;
 use crate::core_ops::{op_cons, op_eq, op_first, op_if, op_listp, op_raise, op_rest};
 use crate::more_ops::{
@@ -10,47 +12,6 @@ use crate::reduction::Response;
 type OpFn<T> = fn(&mut T, <T as Allocator>::Ptr) -> Response<<T as Allocator>::Ptr>;
 
 pub type FLookup<T> = [Option<OpFn<T>>; 256];
-
-pub fn make_f_lookup<T: Allocator>() -> FLookup<T> {
-    let opcode_lookup: [(u8, OpFn<T>); 30] = [
-        (4, op_if),
-        (5, op_cons),
-        (6, op_first),
-        (7, op_rest),
-        (8, op_listp),
-        (9, op_raise),
-        (10, op_eq),
-        (11, op_sha256),
-        (12, op_add),
-        (13, op_subtract),
-        (14, op_multiply),
-        (15, op_divmod),
-        (16, op_substr),
-        (17, op_strlen),
-        (18, op_point_add),
-        (19, op_pubkey_for_exp),
-        (20, op_concat),
-        (22, op_gr),
-        (23, op_gr_bytes),
-        (24, op_logand),
-        (25, op_logior),
-        (26, op_logxor),
-        (27, op_lognot),
-        (28, op_ash),
-        (29, op_lsh),
-        (30, op_not),
-        (31, op_any),
-        (32, op_all),
-        (33, op_softfork),
-        (34, op_div),
-    ];
-    let mut f_lookup: FLookup<T> = [None; 256];
-    for (op, f) in &opcode_lookup {
-        f_lookup[*op as usize] = Some(*f);
-    }
-
-    f_lookup
-}
 
 pub fn opcode_by_name<T: Allocator>(name: &str) -> Option<OpFn<T>> {
     let opcode_lookup: [(&str, OpFn<T>); 30] = [
@@ -93,4 +54,21 @@ pub fn opcode_by_name<T: Allocator>(name: &str) -> Option<OpFn<T>> {
         }
     }
     None
+}
+
+pub fn f_lookup_for_hashmap<A: Allocator>(
+    opcode_lookup_by_name: HashMap<String, &[u8]>,
+) -> FLookup<A> {
+    let mut f_lookup = [None; 256];
+    for (name, idx) in opcode_lookup_by_name.iter() {
+        if idx.len() == 1 {
+            let index = idx[0];
+            let op = opcode_by_name(name);
+            if op.is_none() {
+                panic!("can't find native operator {:?}", name);
+            }
+            f_lookup[index as usize] = op;
+        }
+    }
+    f_lookup
 }
