+++ src/py/native_op_lookup.rs
@@ -9,7 +9,7 @@ use crate::allocator::Allocator;
 use crate::reduction::{EvalErr, Reduction, Response};
 use crate::run_program::OperatorHandler;
 
-use super::f_table::{opcode_by_name, FLookup};
+use super::f_table::{f_lookup_for_hashmap, FLookup};
 
 use super::to_py_node::ToPyNode;
 
@@ -50,17 +50,8 @@ where
         opcode_lookup_by_name: HashMap<String, &[u8]>,
         unknown_op_callback: PyObject,
     ) -> Self {
-        let mut f_lookup = [None; 256];
-        for (name, idx) in opcode_lookup_by_name.iter() {
-            if idx.len() == 1 {
-                let index = idx[0];
-                let op = opcode_by_name(name);
-                if op.is_none() {
-                    panic!("can't find native operator {:?}", name);
-                }
-                f_lookup[index as usize] = op;
-            }
-        }
+        let f_lookup = f_lookup_for_hashmap(opcode_lookup_by_name);
+
         GenericNativeOpLookup {
             py_callback: unknown_op_callback,
             f_lookup,
